New tests and files necessary to fulfil tests
Committing to system. Some tests are commented out. This is because
they are not correctly working yet, but due to timing constraints we
did not have sufficient time to complete all tests. If it were a real
life business scenario, we would have completed all tests to a working
level/